requirements.txt
requests==2.26.0
Django==3.2.7
djangorestframework==3.12.4
django-extensions==3.1.3
flake8==3.9.2
django-cors-headers==3.8.0
